requirements.txt
arxiv
langchainhub
gigachain
gigachain_community
PyMuPDF
rontend
tools
gigagraph
python-docx
unstructured
docx2txt
tiktoken
sentence-transformers
chromadb
pandas==2.0.3
faiss-cpu
duckduckgo-search
gigachain-experimental
telebot
python-dotenv
typing-extensions
